+++ extensions/copymoveto/src/destination-form.tsx
@@ -1,9 +1,22 @@
-import { Action, ActionPanel, Form, LocalStorage, Toast, popToRoot, showToast } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Form,
+  LocalStorage,
+  Toast,
+  popToRoot,
+  showToast,
+  getPreferenceValues,
+} from "@raycast/api";
 
 import { FormValidation, useForm } from "@raycast/utils";
 import { useEffect, useState } from "react";
 import { type Destination, destinationRepo } from "./repo/destination";
 import { checkDirExists } from "./utils/filesystem";
+
+interface Preferences {
+  showHiddenFolders: boolean;
+}
 
 interface EditDestinationProps {
   destination?: Destination;
@@ -13,6 +26,7 @@ export type DestinationForm = Omit<Destination, "directory"> & {
   directory: string[];
 };
 export default function DestinationForm(props: EditDestinationProps) {
+  const preferences = getPreferenceValues<Preferences>();
   const [existingNames, setExistingNames] = useState<string[]>([]);
 
   useEffect(() => {
@@ -100,6 +114,7 @@ export default function DestinationForm(props: EditDestinationProps) {
         allowMultipleSelection={false}
         canChooseDirectories={true}
         canChooseFiles={false}
+        showHiddenFiles={preferences.showHiddenFolders}
         storeValue={true}
         {...itemProps.directory}
       />