+++ extensions/copymoveto/CHANGELOG.md
@@ -1,3 +1,7 @@
 # CopyMoveTo Changelog
 
+## [Hidden Folders Support] - {PR_MERGE_DATE}
+- Added option to show hidden folders in directory picker
+- Added preference setting to toggle hidden folders visibility
+
 ## [Initial Version] - 2025-03-14